+++ classes/class-lsx-customizer-colour.php
@@ -130,10 +130,6 @@ if ( ! class_exists( 'LSX_Customizer_Colour' ) ) {
 			require_once( LSX_CUSTOMIZER_PATH . 'includes/lsx-customizer-colour-deprecated.php' );
 			require_once( LSX_CUSTOMIZER_PATH . 'includes/lsx-customizer-templates.php' );
 
-			if ( class_exists( 'WP_Customize_Control' ) ) {
-				require_once( LSX_CUSTOMIZER_PATH . 'classes/class-lsx-customizer-colour-control.php' );
-			}
-
 			require_once( LSX_CUSTOMIZER_PATH . 'classes/class-lsx-customizer-colour-button.php' );
 			$this->button = new LSX_Customizer_Colour_Button();
 
@@ -185,7 +181,7 @@ if ( ! class_exists( 'LSX_Customizer_Colour' ) ) {
 			 * Colors
 			 */
 			$wp_customize->add_panel( 'colors', array(
-				'title'             => esc_html__( 'Colours', 'lsx-customizer' ),
+				'title'             => esc_html__( 'Site Design', 'lsx-customizer' ),
 				'priority'          => 60,
 			) );